🚚(api) move core base indicators to a dedicated module
To prepare future changes, we decided to move the core base indicator
definition to a dedicated module as more indicator-related files will be
added to this repository.
